Changed the synchronous write to better match node idioms
Node's streams aren't built from a synchronous perspective, so you start running
into timing issues if you try to use them as such. F.e., even though there is a
sync version of `.read` on a stream, if you call it too soon after creating the
stream there won't be anything in the streams buffer and it'll return
`null`/`nil` instead of blocking.

So, for all synchronous work we'll just use the synchronous calls in Node's FS
module and spare ourselves the pain.