Switch to callback async style, and decouple async/sync functions
To avoid forcing library consumers into using core.async, we're going to
adopt a callback style of asynchronicity.

Along with that, I was wrapping the synchronous form of the reader and
writer inside of the async form, but that doesn't make much sense when
node defines sync and async versions of most functions. It had seemed
like a good way to avoid duplicate effort, but it really just
complicated matters.