Add initial walkthrough file
I want to give examples of how to use the abio library, and it seems the
best place to put those examples is with the host implementation. This
is an initial writeup, covering some of the general FS bindings, as well
as how to read a file. I need to add some code showing how to write to a
file as well.